ref(add-bottle): Split photo resolver helpers
Move photo identification helpers and fallback panels out of the resolver component so the add-bottle flow file stays focused on state and orchestration. Behavior is unchanged.
