+++ apps/web/src/components/bottleResolver.tsx
@@ -1,13 +1,11 @@
 "use client";
 
-import type { Inputs, Outputs } from "@peated/server/orpc/router";
 import type { Bottle, BottleRelease } from "@peated/server/types";
 import Button from "@peated/web/components/button";
 import FormError from "@peated/web/components/formError";
 import Header from "@peated/web/components/header";
 import Layout from "@peated/web/components/layout";
 import type { CreateBottlePrefill } from "@peated/web/components/search/createBottleHref";
-import { copyTextToClipboard } from "@peated/web/lib/clipboard";
 import { logError } from "@peated/web/lib/log";
 import { useORPC } from "@peated/web/lib/orpc/context";
 import {
@@ -17,16 +15,12 @@ import {
 } from "@peated/web/lib/orpc/link";
 import { useMutation } from "@tanstack/react-query";
 import {
-  AlertTriangle,
   Camera,
   Check,
   ChevronLeft,
-  Copy,
   ImageIcon,
   Plus,
-  RotateCcw,
   Search,
-  SearchX,
 } from "lucide-react";
 import { useRouter } from "next/navigation";
 import {
@@ -37,7 +31,33 @@ import {
   useState,
 } from "react";
 
-type PhotoIdentification = Outputs["tastings"]["photoIdentification"];
+import {
+  createIdempotencyKey,
+  getAllowedCatalogImageApprovalTarget,
+  getCatalogImageApprovalCopy,
+  getCreateBottlePrefill,
+  getCreateDecision,
+  getCreateNameSeed,
+  getCreateProposalLabel,
+  getManualResultCopy,
+  getMatchedBottleId,
+  getMatchedReleaseId,
+  getProposedName,
+  getSearchSeed,
+  type PhotoIdentification,
+  type PhotoIdentificationCreateInput,
+} from "./bottleResolver.helpers";
+import {
+  EvidencePills,
+  FallbackActions,
+  getPhotoFailureCopyPayload,
+  getPhotoIdentificationCopyPayload,
+  PhotoFailurePanel,
+  type PhotoFailureTrace,
+  PhotoIdentificationTraceFootnote,
+  ResultHeader,
+  SearchBottleCallout,
+} from "./bottleResolverPanels";
 
 export type BottleResolverTarget = {
   bottle: Bottle;
@@ -80,26 +100,6 @@ type Props = {
   enableCatalogImageApproval?: boolean;
 };
 
-type CatalogImageApprovalTarget = "bottle" | "release";
-type ManualResultCopy = {
-  title: string;
-  description: string;
-  createLabel?: string;
-  primaryAction?: "search" | "create";
-};
-type PhotoIdentificationCreateInput =
-  Inputs["tastings"]["photoIdentificationCreate"];
-type PhotoFailureTrace = {
-  traceId: string;
-  file: {
-    name: string;
-    size: number;
-    type: string | null;
-    lastModified: number;
-  };
-  error: string;
-};
-
 const loadingMessages = [
   "Holding it up to the light",
   "Letting the label breathe",
@@ -107,674 +107,6 @@ const loadingMessages = [
   "Asking the tasting room",
   "Comparing the fine print",
 ];
-
-function createIdempotencyKey() {
-  if (typeof crypto !== "undefined" && "randomUUID" in crypto) {
-    return crypto.randomUUID();
-  }
-  return `${Date.now()}-${Math.random().toString(36).slice(2)}`;
-}
-
-function getFieldValue(
-  result: PhotoIdentification | null,
-  field: keyof PhotoIdentification["imageEvidence"]["fieldCandidates"],
-) {
-  const value = result?.imageEvidence.fieldCandidates[field]?.value;
-  if (value === undefined || value === null) return null;
-  if (field === "statedAge") return `${value} years`;
-  if (field === "abv") return `${value}% ABV`;
-  return String(value);
-}
-
-function getRawFieldValue(
-  result: PhotoIdentification | null,
-  field: keyof PhotoIdentification["imageEvidence"]["fieldCandidates"],
-) {
-  const value = result?.imageEvidence.fieldCandidates[field]?.value;
-  if (value === undefined || value === null || value === "") return null;
-  return value;
-}
-
-function getRawStringFieldValue(
-  result: PhotoIdentification | null,
-  field: keyof PhotoIdentification["imageEvidence"]["fieldCandidates"],
-) {
-  const value = getRawFieldValue(result, field);
-  return typeof value === "string" ? value : null;
-}
-
-function getRawNumberFieldValue(
-  result: PhotoIdentification | null,
-  field: keyof PhotoIdentification["imageEvidence"]["fieldCandidates"],
-) {
-  const value = getRawFieldValue(result, field);
-  return typeof value === "number" ? value : null;
-}
-
-function getSearchSeed(result: PhotoIdentification | null) {
-  const brand = getFieldValue(result, "brand");
-  const expression = getFieldValue(result, "expression");
-  return [brand, expression].filter(Boolean).join(" ");
-}
-
-function getCreateNameSeed(result: PhotoIdentification | null) {
-  const decision = getCreateDecisionLike(result);
-  return (
-    getRawStringFieldValue(result, "expression") ??
-    getProposedBottle(decision)?.name ??
-    ""
-  );
-}
-
-function getCreateBottlePrefill(
-  result: PhotoIdentification | null,
-): CreateBottlePrefill {
-  const decision = getCreateDecisionLike(result);
-  const proposedBottle = getProposedBottle(decision);
-  const proposedRelease = getProposedRelease(decision);
-
-  return {
-    brandName:
-      getRawStringFieldValue(result, "brand") ??
-      proposedBottle?.brand.name ??
-      null,
-    statedAge: getRawNumberFieldValue(result, "statedAge"),
-    abv: getRawNumberFieldValue(result, "abv"),
-    edition:
-      getRawStringFieldValue(result, "edition") ??
-      proposedRelease?.edition ??
-      null,
-    vintageYear: getRawNumberFieldValue(result, "vintageYear"),
-    releaseYear: getRawNumberFieldValue(result, "releaseYear"),
-  };
-}
-
-function hasRecognizedLabelDetails(result: PhotoIdentification | null) {
-  return Boolean(
-    getFieldValue(result, "brand") ||
-    getFieldValue(result, "expression") ||
-    getFieldValue(result, "statedAge") ||
-    getFieldValue(result, "abv") ||
-    getFieldValue(result, "edition") ||
-    getFieldValue(result, "vintageYear") ||
-    getFieldValue(result, "releaseYear") ||
-    getFieldValue(result, "caskNumber"),
-  );
-}
-
-function getMatchedBottleId(result: PhotoIdentification | null) {
-  if (
-    result?.suggestedNextStep === "confirm_match" &&
-    result.classification.status === "classified" &&
-    result.classification.decision.action === "match"
-  ) {
-    return result.classification.decision.matchedBottleId;
-  }
-  return null;
-}
-
-function getMatchedReleaseId(result: PhotoIdentification | null) {
-  if (
-    result?.classification.status === "classified" &&
-    result.classification.decision.action === "match"
-  ) {
-    return result.classification.decision.matchedReleaseId;
-  }
-  return null;
-}
-
-function getCreateDecision(result: PhotoIdentification | null) {
-  if (
-    result?.suggestedNextStep !== "confirm_create" ||
-    result.classification.status !== "classified"
-  ) {
-    return null;
-  }
-
-  return getCreateDecisionLike(result);
-}
-
-function getCreateDecisionLike(result: PhotoIdentification | null) {
-  if (result?.classification.status !== "classified") {
-    return null;
-  }
-
-  switch (result.classification.decision.action) {
-    case "create_bottle":
-    case "create_release":
-    case "create_bottle_and_release":
-      return result.classification.decision;
-    default:
-      return null;
-  }
-}
-
-type CreateDecisionLike = NonNullable<ReturnType<typeof getCreateDecisionLike>>;
-
-function getProposedBottle(decision: CreateDecisionLike | null) {
-  if (!decision) return null;
-  if (
-    decision.action === "create_bottle" ||
-    decision.action === "create_bottle_and_release"
-  ) {
-    return decision.proposedBottle;
-  }
-  return null;
-}
-
-function getProposedRelease(decision: CreateDecisionLike | null) {
-  if (!decision) return null;
-  if (
-    decision.action === "create_release" ||
-    decision.action === "create_bottle_and_release"
-  ) {
-    return decision.proposedRelease;
-  }
-  return null;
-}
-
-function getProposedName(result: PhotoIdentification | null) {
-  const decision = getCreateDecision(result);
-  if (!decision) return null;
-
-  if (decision.action === "create_release") {
-    return decision.proposedRelease.edition ?? "New release";
-  }
-
-  if (decision.action === "create_bottle_and_release") {
-    return [
-      decision.proposedBottle.brand.name,
-      decision.proposedBottle.name,
-      decision.proposedRelease.edition,
-    ]
-      .filter(Boolean)
-      .join(" ");
-  }
-
-  return [decision.proposedBottle.brand.name, decision.proposedBottle.name]
-    .filter(Boolean)
-    .join(" ");
-}
-
-function getParentBottleName(result: PhotoIdentification | null) {
-  const decision = getCreateDecision(result);
-  if (!decision || decision.action !== "create_release") return null;
-
-  const parent = result?.classification.artifacts.candidates.find(
-    (candidate) =>
-      candidate.bottleId === decision.parentBottleId &&
-      (candidate.releaseId === null || candidate.releaseId === undefined),
-  );
-  return parent?.bottleFullName || parent?.fullName || null;
-}
-
-function getCreateProposalLabel(result: PhotoIdentification | null) {
-  const decision = getCreateDecision(result);
-  if (!decision) return null;
-  const parentBottleName = getParentBottleName(result);
-
-  if (decision.action === "create_release") {
-    return {
-      title: "Bottling not in Peated",
-      description: parentBottleName
-        ? `Create a new bottling for ${parentBottleName}.`
-        : "Create a new bottling for this bottle.",
-    };
-  }
-
-  if (decision.action === "create_bottle_and_release") {
-    return {
-      title: "Bottle not in Peated",
-      description: "Create the bottle and its specific bottling.",
-    };
-  }
-
-  return {
-    title: "Bottle not in Peated",
-    description: "Create a new bottle from this label.",
-  };
-}
-
-function getCatalogImageApprovalTarget(
-  action: "create_bottle" | "create_release" | "create_bottle_and_release",
-): CatalogImageApprovalTarget {
-  return action === "create_bottle" ? "bottle" : "release";
-}
-
-function getAllowedCatalogImageApprovalTarget(
-  result: PhotoIdentification,
-  enabled: boolean,
-) {
-  const decision = getCreateDecision(result);
-  if (
-    !enabled ||
-    !decision ||
-    result.imageEvidence.photoSuitability.suitableAsBottleImage !== true
-  ) {
-    return null;
-  }
-
-  return getCatalogImageApprovalTarget(decision.action);
-}
-
-function getCatalogImageApprovalCopy(target: CatalogImageApprovalTarget) {
-  if (target === "bottle") {
-    return {
-      label: "Set as Bottle Image",
-      help: "This photo will be shown as the public image for the new bottle.",
-    };
-  }
-
-  return {
-    label: "Set as Release Image",
-    help: "This photo will be shown as the public image for the new release.",
-  };
-}
-
-function getManualResultCopy(
-  result: PhotoIdentification | null,
-): ManualResultCopy {
-  const action =
-    result?.classification.status === "classified"
-      ? result.classification.decision.action
-      : null;
-
-  if (result?.suggestedNextStep === "needs_review") {
-    return {
-      title: "We couldn't identify this bottle",
-      description:
-        "We found a possible match, but it was not reliable enough to use automatically. Search can still find the right bottle.",
-      createLabel: undefined,
-    };
-  }
-
-  if (action === "match") {
-    return {
-      title: "We couldn't confirm the match",
-      description:
-        "We found a possible match, but it was not reliable enough to use automatically.",
-      createLabel: undefined,
-    };
-  }
-
-  if (action === "no_match") {
-    if (hasRecognizedLabelDetails(result)) {
-      return {
-        title: "We couldn't find this bottle",
-        description:
-          "We found label details, but not enough to choose an existing bottle automatically. Review them before creating it in Peated.",
-        createLabel: "Create Bottle",
-        primaryAction: "create" as const,
-      };
-    }
-
-    return {
-      title: "We couldn't identify the bottle",
-      description:
-        "Search can still find it, or you can start over with a clearer photo.",
-      createLabel: "Create Manually",
-      primaryAction: "search" as const,
-    };
-  }
-
-  if (
-    action === "create_bottle" ||
-    action === "create_release" ||
-    action === "create_bottle_and_release"
-  ) {
-    return {
-      title: "We couldn't find this bottle",
-      description:
-        "We found label details, but not enough to choose an existing bottle automatically. Review them before creating it in Peated.",
-      createLabel: "Create Bottle",
-      primaryAction: "create" as const,
-    };
-  }
-
-  return {
-    title: "We couldn't identify the bottle",
-    description:
-      "Search can still find it, or you can start over with another photo.",
-    createLabel: "Create Manually",
-    primaryAction: "search" as const,
-  };
-}
-
-function EvidencePills({ result }: { result: PhotoIdentification | null }) {
-  const fields = [
-    ["Brand", getFieldValue(result, "brand")],
-    ["Expression", getFieldValue(result, "expression")],
-    ["Age", getFieldValue(result, "statedAge")],
-    ["ABV", getFieldValue(result, "abv")],
-    ["Edition", getFieldValue(result, "edition")],
-    ["Vintage", getFieldValue(result, "vintageYear")],
-    ["Release", getFieldValue(result, "releaseYear")],
-    ["Cask", getFieldValue(result, "caskNumber")],
-  ].filter(([, value]) => value);
-
-  if (!fields.length) return null;
-
-  return (
-    <div className="flex flex-wrap gap-2">
-      {fields.map(([label, value]) => (
-        <div
-          key={label}
-          className="rounded border border-slate-800 bg-slate-950 px-3 py-2 text-sm"
-        >
-          <span className="text-muted">{label}</span>{" "}
-          <span className="font-medium text-white">{value}</span>
-        </div>
-      ))}
-    </div>
-  );
-}
-
-function ResultHeader({
-  previewUrl,
-  icon,
-  title,
-  description,
-  children,
-}: {
-  previewUrl: string | null;
-  icon: ReactNode;
-  title: string;
-  description: string;
-  children?: ReactNode;
-}) {
-  return (
-    <div className="flex items-start gap-4">
-      {previewUrl && (
-        <img
-          src={previewUrl}
-          alt=""
-          className="h-16 w-16 shrink-0 rounded object-cover sm:h-[96px] sm:w-[96px]"
-        />
-      )}
-      <div className="min-w-0 flex-1 space-y-3">
-        <div className="flex items-start gap-3">
-          {icon}
-          <div>
-            <div className="font-semibold text-white">{title}</div>
-            <div className="text-muted mt-1 text-sm">{description}</div>
-          </div>
-        </div>
-        {children}
-      </div>
-    </div>
-  );
-}
-
-function OrDivider() {
-  return (
-    <div className="flex items-center gap-3">
-      <div className="h-px flex-1 bg-slate-800" />
-      <div className="text-muted text-xs font-semibold uppercase tracking-wide">
-        Or
-      </div>
-      <div className="h-px flex-1 bg-slate-800" />
-    </div>
-  );
-}
-
-function PhotoFailurePanel({
-  previewUrl,
-  title,
-  description,
-  searchHref,
-  searchLabel,
-  createBottleHref,
-  createBottleLabel = "Create Manually",
-  primaryAction = "search",
-  onStartOver,
-  variant,
-  children,
-}: {
-  previewUrl: string | null;
-  title: string;
-  description: string;
-  searchHref: string;
-  searchLabel: string;
-  createBottleHref?: string | null;
-  createBottleLabel?: string;
-  primaryAction?: "search" | "create";
-  onStartOver: () => void;
-  variant: "error" | "no-match";
-  children?: ReactNode;
-}) {
-  const isError = variant === "error";
-
-  return (
-    <div
-      className={`rounded border p-4 lg:p-6 ${
-        isError
-          ? "border-red-900/70 bg-red-950/20"
-          : "border-slate-800 bg-slate-950/50"
-      }`}
-    >
-      <div className="space-y-5">
-        <div className="flex flex-col gap-4 sm:flex-row sm:items-start">
-          {previewUrl && (
-            <img
-              src={previewUrl}
-              alt="Selected bottle label"
-              className="h-24 w-24 rounded object-cover"
-            />
-          )}
-          <div className="min-w-0 flex-1 space-y-4">
-            <div className="flex items-start gap-3">
-              <div
-                className={`rounded-full p-2 ${
-                  isError
-                    ? "bg-red-900/60 text-red-100"
-                    : "text-highlight bg-slate-800"
-                }`}
-              >
-                {isError ? (
-                  <AlertTriangle className="h-5 w-5" />
-                ) : (
-                  <SearchX className="h-5 w-5" />
-                )}
-              </div>
-              <div>
-                <div className="font-semibold text-white">{title}</div>
-                <div className="text-muted mt-1 text-sm">{description}</div>
-              </div>
-            </div>
-            {children}
-          </div>
-        </div>
-        <div
-          className={`grid w-full gap-3 ${
-            createBottleHref ? "sm:grid-cols-3" : "sm:grid-cols-2"
-          }`}
-        >
-          {primaryAction === "create" && createBottleHref ? (
-            <>
-              <Button
-                href={createBottleHref}
-                color="highlight"
-                fullWidth
-                icon={<Plus className="h-4 w-4" />}
-              >
-                {createBottleLabel}
-              </Button>
-              <Button
-                href={searchHref}
-                fullWidth
-                icon={<Search className="h-4 w-4" />}
-              >
-                {searchLabel}
-              </Button>
-            </>
-          ) : (
-            <>
-              <Button
-                href={searchHref}
-                color="highlight"
-                fullWidth
-                icon={<Search className="h-4 w-4" />}
-              >
-                {searchLabel}
-              </Button>
-              {createBottleHref && (
-                <Button
-                  href={createBottleHref}
-                  fullWidth
-                  icon={<Plus className="h-4 w-4" />}
-                >
-                  {createBottleLabel}
-                </Button>
-              )}
-            </>
-          )}
-          <Button
-            fullWidth
-            onClick={onStartOver}
-            icon={<RotateCcw className="h-4 w-4" />}
-          >
-            Start Over
-          </Button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function getPhotoIdentificationCopyPayload(
-  result: PhotoIdentification,
-  traceId: string,
-) {
-  return JSON.stringify(
-    {
-      traceId,
-      pendingImage: result.pendingImage,
-      suggestedNextStep: result.suggestedNextStep,
-      diagnostics: result.diagnostics,
-      imageEvidence: result.imageEvidence,
-      classification: result.classification,
-    },
-    null,
-    2,
-  );
-}
-
-function PhotoIdentificationTraceFootnote({
-  traceId,
-  copyPayload,
-}: {
-  traceId: string;
-  copyPayload: string;
-}) {
-  const [copied, setCopied] = useState(false);
-  const [copyFailed, setCopyFailed] = useState(false);
-
-  async function copyTracePayload() {
-    setCopied(false);
-    setCopyFailed(false);
-
-    try {
-      await copyTextToClipboard(copyPayload);
-      setCopied(true);
-      window.setTimeout(() => setCopied(false), 1800);
-    } catch (err) {
-      logError(err);
-      setCopyFailed(true);
-    }
-  }
-
-  return (
-    <div className="text-muted mt-4 flex items-center gap-1.5 px-1 text-xs">
-      <span className="min-w-0 truncate">Trace ID: {traceId}</span>
-      <button
-        type="button"
-        className="focus:ring-highlight shrink-0 rounded p-1 transition hover:bg-slate-800 hover:text-white focus:outline-none focus:ring-2"
-        onClick={() => void copyTracePayload()}
-        title="Copy photo identification payload"
-        aria-label="Copy photo identification payload"
-      >
-        <Copy className="h-3.5 w-3.5" />
-      </button>
-      <span className="sr-only" role="status">
-        {copied
-          ? "Photo identification payload copied"
-          : copyFailed
-            ? "Photo identification payload copy failed"
-            : ""}
-      </span>
-    </div>
-  );
-}
-
-function getPhotoFailureCopyPayload(trace: PhotoFailureTrace) {
-  return JSON.stringify(
-    {
-      traceId: trace.traceId,
-      file: trace.file,
-      error: trace.error,
-      context: "add_bottle_photo_identification",
-      rpc: "tastings.photoIdentification",
-    },
-    null,
-    2,
-  );
-}
-
-function FallbackActions({
-  searchHref,
-  searchLabel,
-  onStartOver,
-  showStartOver = false,
-}: {
-  searchHref: string;
-  searchLabel: string;
-  onStartOver?: () => void;
-  showStartOver?: boolean;
-}) {
-  return (
-    <div className="space-y-3">
-      <OrDivider />
-      <div className="mx-auto grid w-full gap-3 sm:w-1/2 sm:grid-cols-2">
-        <Button
-          href={searchHref}
-          fullWidth
-          icon={<Search className="h-4 w-4" />}
-        >
-          {searchLabel}
-        </Button>
-        {showStartOver && onStartOver && (
-          <Button
-            fullWidth
-            onClick={onStartOver}
-            icon={<RotateCcw className="h-4 w-4" />}
-          >
-            Start Over
-          </Button>
-        )}
-      </div>
-    </div>
-  );
-}
-
-function SearchBottleCallout({ searchHref }: { searchHref: string }) {
-  return (
-    <section className="mx-3 rounded border border-slate-800 bg-slate-950/50 p-4 sm:mx-0">
-      <div className="flex flex-col gap-3 sm:flex-row sm:items-center sm:justify-between">
-        <div>
-          <div className="font-semibold text-white">Prefer to search?</div>
-          <div className="text-muted mt-1 text-sm">
-            Find a bottle in Peated or start a new one manually.
-          </div>
-        </div>
-        <div className="shrink-0">
-          <Button href={searchHref} icon={<Search className="h-4 w-4" />}>
-            Search Bottles
-          </Button>
-        </div>
-      </div>
-    </section>
-  );
-}
 
 export default function BottleResolver({
   onResolve,
